Check file syntax on save
This prevents running perltidy on non-perl files (Which i thought didn't happen based on the extension config, but was confused) and will allow for it to work with multiple perl syntax and not just the default perl syntax shipped with nova.

I'm going to add a new tree sitter based syntax which needs to have a different name from the one shipped with Nova.